added navbar, vue router
