+++ README.md
@@ -1,4 +1,4 @@
 # notes-app
-## This a simple app can create a new notes, edit existing one, and ofc delete it.
+## This is a simple app can create a new notes, edit existing ones, and ofc delete it.
 ## Link: https://vladislavsaleev.github.io/notes-app/
 ### All changes will be stored in localStorage, but i'm not recommend to keep here a important things, because it easy to erase
