+++ src/main.js
@@ -1,4 +1,10 @@
 import { createApp } from 'vue'
 import App from './App.vue'
+import router from '@/router/router'
 
-createApp(App).mount('#app')
+const app = createApp(App)
+
+app
+    .use(router)
+    .mount('#app')
+
