+++ src/pages/About.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="about">Something about this APP.</div>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style lang="scss" scoped>
+.about {
+  margin-top: 70px;
+  text-align: center;
+}
+</style>
